refactor: parallel-route + 컴포넌트 분리
