+++ src/app/(routes)/(tabs)/note/(create)/@header/default.tsx
