+++ src/app/(routes)/(tabs)/note/(create)/notion/page.tsx
@@ -1,4 +1,3 @@
-import Header from '../ui/header'
 import CreateQuizButton from '@/features/note/components/create-quiz-button'
 import { CreateNoteProvider } from '@/features/note/contexts/create-note-context'
 import Icon from '@/shared/components/custom/icon'
@@ -7,8 +6,6 @@ import Text from '@/shared/components/ui/text'
 const CreateWithNotionPage = () => {
   return (
     <>
-      <Header />
-
       <CreateNoteProvider>
         <div className="fixed top-[54px] flex w-full max-w-mobile items-center justify-between bg-background-base-02 px-[16px] py-[11px]">
           <div className="flex items-center">
