+++ src/app/(routes)/(tabs)/note/(main)/@header/default.tsx
