+++ src/app/(routes)/(tabs)/note/(main)/layout.tsx
@@ -1,6 +1,6 @@
 import { FunctionComponent, PropsWithChildren } from 'react'
 import type { Metadata } from 'next'
-import { InquiryProvider } from '@/features/user/contexts/inquiry-context'
+import { QuizNoteProvider } from '@/features/note/contexts/quiz-note-context'
 
 export const metadata: Metadata = {}
 
@@ -10,10 +10,12 @@ interface LayoutProps extends PropsWithChildren {
 
 const Layout: FunctionComponent<LayoutProps> = ({ header, children }) => {
   return (
-    <InquiryProvider>
-      {header}
-      {children}
-    </InquiryProvider>
+    <div className="flex h-[calc(100dvh-88px)] w-full flex-col overflow-hidden bg-background-base-02 text-text-primary">
+      <QuizNoteProvider>
+        {header}
+        {children}
+      </QuizNoteProvider>
+    </div>
   )
 }
