+++ src/app/(routes)/(tabs)/note/[id]/(main)/layout.tsx
@@ -1,6 +1,5 @@
 import { FunctionComponent, PropsWithChildren } from 'react'
 import type { Metadata } from 'next'
-import { InquiryProvider } from '@/features/user/contexts/inquiry-context'
 
 export const metadata: Metadata = {}
 
@@ -10,10 +9,10 @@ interface LayoutProps extends PropsWithChildren {
 
 const Layout: FunctionComponent<LayoutProps> = ({ header, children }) => {
   return (
-    <InquiryProvider>
+    <>
       {header}
       {children}
-    </InquiryProvider>
+    </>
   )
 }
