+++ src/app/(routes)/(tabs)/profile/(main)/layout.tsx
@@ -1,12 +1,19 @@
 import { FunctionComponent, PropsWithChildren } from 'react'
 import type { Metadata } from 'next'
+import Header from '@/features/user/header'
 
 export const metadata: Metadata = {}
 
 interface LayoutProps extends PropsWithChildren {}
 
 const Layout: FunctionComponent<LayoutProps> = ({ children }) => {
-  return <main>{children}</main>
+  return (
+    <>
+      <Header />
+
+      {children}
+    </>
+  )
 }
 
 export default Layout
