+++ src/app/(routes)/(tabs)/profile/inquiry/index.stories.tsx
@@ -1,5 +1,7 @@
 import { Meta, StoryObj } from '@storybook/react'
 import InquiryPage from './page'
+import Header from './@header/default'
+import { InquiryProvider } from '@/features/user/contexts/inquiry-context'
 
 const meta: Meta<typeof InquiryPage> = {
   title: 'Page/Inquiry',
@@ -10,7 +12,9 @@ const meta: Meta<typeof InquiryPage> = {
   decorators: [
     (Story) => (
       <div className="mx-auto max-w-mobile">
-        <Story />
+        <InquiryProvider>
+          <Story />
+        </InquiryProvider>
       </div>
     ),
   ],
@@ -21,5 +25,10 @@ export default meta
 type Story = StoryObj<typeof InquiryPage>
 
 export const Default: Story = {
-  render: () => <InquiryPage />,
+  render: () => (
+    <>
+      <Header />
+      <InquiryPage />
+    </>
+  ),
 }
