+++ src/app/(routes)/(tabs)/profile/notification/index.stories.tsx
@@ -1,5 +1,6 @@
 import { Meta, StoryFn, StoryObj } from '@storybook/react'
 import NotificationPage from './page'
+import Layout from './layout'
 
 const meta: Meta<typeof NotificationPage> = {
   title: 'Page/Notification',
@@ -18,4 +19,10 @@ const meta: Meta<typeof NotificationPage> = {
 
 export default meta
 
-export const Default: StoryObj<typeof NotificationPage> = {}
+export const Default: StoryObj<typeof NotificationPage> = {
+  render: () => (
+    <Layout>
+      <NotificationPage />
+    </Layout>
+  ),
+}
