+++ src/app/(routes)/(tabs)/profile/verify-email/index.stories.tsx
@@ -1,5 +1,6 @@
 import { Meta, StoryObj } from '@storybook/react'
 import EmailPage from './page'
+import Layout from './layout'
 
 const meta: Meta<typeof EmailPage> = {
   title: 'Page/Email',
@@ -21,5 +22,9 @@ export default meta
 type Story = StoryObj<typeof EmailPage>
 
 export const Default: Story = {
-  render: () => <EmailPage />,
+  render: () => (
+    <Layout>
+      <EmailPage />
+    </Layout>
+  ),
 }
